Add files via upload
 A computer's Central Processing Unit (CPU) built on a single Integrated Circuit (IC) is called a microprocessor.
It is a programmable, multipurpose, clock-driven, register-based electronic device that reads binary instructions from memory storage, accepts binary data as input and processes data according to those instructions, and provides results as output.
A microprocessor consists of an ALU, control unit, and register array. Where ALU performs arithmetic and logical operations on the data received from an input device or memory. The Control unit controls the instructions and flow of data within the computer. And, register array consists of registers identified by letters like B, C, D, E, H, L, and accumulator.
Instructions in 8086:
The instruction set is the group of commands that the microprocessor can understand. It is an interface between hardware and software.
some programs are compiled here.